+++ src/main/java/callprotector/spring/domain/calllog/entity/CallLog.java
@@ -1,6 +1,6 @@
 package callprotector.spring.domain.calllog.entity;
 
-import callprotector.spring.domain.CallSession;
+import callprotector.spring.domain.callsession.entity.CallSession;
 import callprotector.spring.global.common.entity.BaseEntity;
 import callprotector.spring.global.common.enums.CallTrack;
 import jakarta.persistence.*;
