+++ src/main/java/callprotector/spring/domain/calllog/repository/CallLogRepository.java
@@ -1,7 +1,7 @@
 package callprotector.spring.domain.calllog.repository;
 
 import callprotector.spring.domain.calllog.entity.CallLog;
-import callprotector.spring.domain.CallSession;
+import callprotector.spring.domain.callsession.entity.CallSession;
 import callprotector.spring.global.common.enums.CallTrack;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
