+++ src/main/java/callprotector/spring/domain/calllog/service/CallLogServiceImpl.java
@@ -1,11 +1,11 @@
 package callprotector.spring.domain.calllog.service;
 
 import callprotector.spring.domain.calllog.entity.CallLog;
-import callprotector.spring.domain.CallSession;
+import callprotector.spring.domain.callsession.entity.CallSession;
 import callprotector.spring.domain.calllog.repository.CallLogRepository;
 import callprotector.spring.global.common.enums.CallTrack;
 import callprotector.spring.domain.abuse.service.AbuseService;
-import callprotector.spring.service.CallSessionService.CallSessionService;
+import callprotector.spring.domain.callsession.service.CallSessionService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
