+++ src/main/java/callprotector/spring/domain/callsession/dto/request/CallSessionRequestDTO.java
@@ -1,4 +1,4 @@
-package callprotector.spring.web.dto.request;
+package callprotector.spring.domain.callsession.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
