+++ src/main/java/callprotector/spring/domain/callsession/repository/CallSessionRepositoryImpl.java
@@ -1,6 +1,6 @@
-package callprotector.spring.repository;
+package callprotector.spring.domain.callsession.repository;
 
-import callprotector.spring.domain.CallSession;
+import callprotector.spring.domain.callsession.entity.CallSession;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.TypedQuery;
