+++ src/main/java/callprotector/spring/global/handler/SttContext.java
@@ -15,7 +15,7 @@ import callprotector.spring.global.client.FastClient;
 import callprotector.spring.domain.CallSttLog;
 import callprotector.spring.global.common.enums.CallTrack;
 import callprotector.spring.domain.calllog.service.CallLogService;
-import callprotector.spring.service.CallSessionService.CallSessionService;
+import callprotector.spring.domain.callsession.service.CallSessionService;
 import callprotector.spring.service.CallSttLogService.CallSttLogService;
 import callprotector.spring.web.dto.response.CallSttLogResponseDTO;
