+++ src/main/java/callprotector/spring/global/handler/TwilioMediaStreamProcessor.java
@@ -14,11 +14,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import callprotector.spring.global.client.FastClient;
 import callprotector.spring.global.common.enums.CallTrack;
 import callprotector.spring.domain.calllog.service.CallLogService;
-import callprotector.spring.service.CallSessionService.CallSessionService;
+import callprotector.spring.domain.callsession.service.CallSessionService;
 import callprotector.spring.service.CallSttLogService.CallSttLogService;
 import callprotector.spring.domain.user.service.UserService;
-import callprotector.spring.web.dto.request.CallSessionRequestDTO;
-import callprotector.spring.web.dto.response.CallSessionResponseDTO;
+import callprotector.spring.domain.callsession.dto.request.CallSessionRequestDTO;
+import callprotector.spring.domain.callsession.dto.response.CallSessionResponseDTO;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
