+++ src/main/java/callprotector/spring/global/handler/TwilioMediaStreamsHandler.java
@@ -2,7 +2,7 @@ package callprotector.spring.global.handler;
 
 import callprotector.spring.global.client.FastClient;
 import callprotector.spring.domain.calllog.service.CallLogService;
-import callprotector.spring.service.CallSessionService.CallSessionService;
+import callprotector.spring.domain.callsession.service.CallSessionService;
 import callprotector.spring.service.CallSttLogService.CallSttLogService;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
